+++ bundles/mapping/mapmodule/plugin/zoombar/Portti2Zoombar.js
@@ -30,6 +30,18 @@ Oskari.clazz.define('Oskari.mapframework.bundle.mapmodule.plugin.Portti2Zoombar'
         this._name = 'Portti2Zoombar';
         this._slider = null;
         this._suppressEvents = false;
+        this._desktopStyles = {
+            plus: {
+                css: {}
+            },
+            minus: {
+                css: {}
+            }
+        };
+
+        this._mobileDefs = {
+            height: 500
+        };
 
         this.toolStyles = {
             'default': {
@@ -223,8 +235,47 @@ Oskari.clazz.define('Oskari.mapframework.bundle.mapmodule.plugin.Portti2Zoombar'
             return {
                 AfterMapMoveEvent: function (event) {
                     me._setZoombarValue(event.getZoom());
+                },
+                MapSizeChangedEvent: function (evt) {
+                    me._handleMapSizeChanges({width:evt.getWidth(), height:evt.getHeight()});
                 }
             };
+        },
+
+        /**
+         * @method  @private _handleMapSizeChanges handle map size changes
+         * @param  {Object} size {width:100, height:200}
+         */
+        _handleMapSizeChanges: function(size, el){
+            var me = this,
+                div = el || this.getElement(),
+                plus = div.find('.pzbDiv-plus'),
+                minus = div.find('.pzbDiv-minus'),
+                slider = div.find('div.slider');
+
+            if(size.height < me._mobileDefs.height) {
+                slider.hide();
+                plus.css({
+                    'background-image': 'url("' + this.getImagePath() + 'zoombar_plus_mobile.png")',
+                    'width': 43,
+                    'height': 43
+                });
+
+                minus.css({
+                    'background-image': 'url("' + this.getImagePath() + 'zoombar_minus_mobile.png")',
+                    'width': 43,
+                    'height': 43
+                });
+                div.width(43);
+
+            }
+            else {
+                slider.show();
+                plus.css(me._desktopStyles.plus.css);
+                minus.css(me._desktopStyles.minus.css);
+                div.width(18);
+            }
+
         },
 
         _setLayerToolsEditModeImpl: function () {
@@ -246,6 +297,7 @@ Oskari.clazz.define('Oskari.mapframework.bundle.mapmodule.plugin.Portti2Zoombar'
          *
          */
         changeToolStyle: function (style, div) {
+            var me = this;
             // FIXME move under _setStyle or smthn...
             div = div || this.getElement();
 
@@ -268,7 +320,8 @@ Oskari.clazz.define('Oskari.mapframework.bundle.mapmodule.plugin.Portti2Zoombar'
                 cursor = div.find('.ui-slider-handle'),
                 plus = div.find('.pzbDiv-plus'),
                 minus = div.find('.pzbDiv-minus'),
-                slider = div.find('div.slider');
+                slider = div.find('div.slider'),
+                mapModule = me.getMapModule();
 
             // FIXME get rid of this, rounded style should be fixed instead
             // Used to get the cursor to the right position since
@@ -288,16 +341,26 @@ Oskari.clazz.define('Oskari.mapframework.bundle.mapmodule.plugin.Portti2Zoombar'
                     'height': '',
                     'margin-left': ''
                 });
-                plus.css({
-                    'background-image': '',
-                    'width': '',
-                    'height': ''
-                });
-                minus.css({
-                    'background-image': '',
-                    'width': '',
-                    'height': ''
-                });
+
+                me._desktopStyles = {
+                    plus: {
+                        css: {
+                            'background-image': '',
+                            'width': '',
+                            'height': ''
+                        }
+                    },
+                    minus: {
+                        css: {
+                            'background-image': '',
+                            'width': '',
+                            'height': ''
+                        }
+                    }
+                };
+                plus.css(me._desktopStyles.plus.css);
+                minus.css(me._desktopStyles.minus.css);
+
                 slider.css({
                     'height': sliderHeight + 'px'
                 });
@@ -313,20 +376,32 @@ Oskari.clazz.define('Oskari.mapframework.bundle.mapmodule.plugin.Portti2Zoombar'
                     'height': style.heightCursor,
                     'margin-left': (isRounded ? '2px' : '0')
                 });
-                plus.css({
-                    'background-image': 'url("' + zoombarPlusImg + '")',
-                    'width': style.widthPlus,
-                    'height': style.heightPlus
-                });
-                minus.css({
-                    'background-image': 'url("' + zoombarMinusImg + '")',
-                    'width': style.widthMinus,
-                    'height': style.heightMinus
-                });
+
+                me._desktopStyles = {
+                    plus: {
+                        css: {
+                            'background-image': 'url("' + zoombarPlusImg + '")',
+                            'width': style.widthPlus,
+                            'height': style.heightPlus
+                        }
+                    },
+                    minus: {
+                        css: {
+                            'background-image': 'url("' + zoombarMinusImg + '")',
+                            'width': style.widthMinus,
+                            'height': style.heightMinus
+                        }
+                    }
+                };
+
+                plus.css(me._desktopStyles.plus.css);
+                minus.css(me._desktopStyles.minus.css);
                 slider.css({
                     'height': sliderHeight + 'px'
                 });
             }
+
+            me._handleMapSizeChanges(mapModule.getSize(), div);
         }
     }, {
         'extend': ['Oskari.mapping.mapmodule.plugin.BasicMapModulePlugin'],